Move DeleteAccountModalContent to its own folder
